requirements.txt
pandas>=1.5.0
datasets>=2.12.0
huggingface_hub>=0.14.0
transformers>=4.28.0
sentence-transformers>=2.2.0
torch>=2.0.0
python-dotenv>=0.21.0
openai>=1.0.0
regex>=2022.10.31
sacrebleu>=2.0.0
rouge-score>=0.1.2
bert-score>=0.3.13